fix: new hook
